LPS-106037 Add Question component
